Refactor: [사용자 관리] Kakao 응답 어댑터 적용 (MOYEO-27)
- 어댑터 패턴 적용
